# api-spec.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: >-
      https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project
  version: 3.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/businesses/{id}/inventory/':
    get:
      tags:
        - 'U19: Create inventory'
        - Business Accounts
        - Inventory
      summary: Retrieve a business's product inventory
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
        - name: page
          in: query
          description: The page number to retrieve (1 indexed)
          schema:
            type: integer
        - name: resultsPerPage
          in: query
          description: The number of results to show per page
          schema:
            type: integer
        - name: orderBy
          in: query
          description: Card property to order results by
          schema:
            type: string
            enum:
              - name
              - description
              - manufacturer
              - recommendedRetailPrice
              - created
              - quantity
              - pricePerItem
              - totalPrice
              - manufactured
              - sellBy
              - bestBefore
              - expires
              - productCode
        - name: reverse
          in: query
          description: Whether to flip default sort order (default is ascending)
          schema:
            type: boolean
      responses:
        '200':
          description: Inventory retrieved successfully. This could be an empty array.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required:
                      - results
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/InventoryItem'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            The account performing the request is neither an administrator of
            the business, nor a global application admin.
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    post:
      tags:
        - 'U19: Create inventory'
        - Business Accounts
        - Inventory
      summary: Add a new inventory item to a business's inventory
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Inventory item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  inventoryItemId:
                    type: integer
                    description: The ID of the inventory item that has just been created
                    example: 101
        '400':
          description: >-
            There was some error with the data supplied by the user, appropriate
            error message(s) should be shown to user
        '403':
          description: >-
            Forbidden: Returned when a user tries to add an inventory item to
            the inventory of a business they do not administer AND the user is
            not a global application admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
  '/businesses/{businessId}/inventory/{inventoryItemId}':
    put:
      tags:
        - 'U20: Modify inventory entries'
        - Business Accounts
        - Inventory
      summary: Modify an existing inventory item
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business to search catalogue of
          required: true
          schema:
            type: integer
        - name: inventoryItemId
          in: path
          description: ID of inventory item to update
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        '200':
          description: Inventory item updated successfully
        '400':
          description: >-
            There was some error with the data supplied by the user, appropriate
            error message(s) should be shown to user
        '403':
          description: >-
            Forbidden: Returned when a user tries to update an inventory item
            for a business they do not administer AND the user is not a global
            application admin
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
...
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: >-
        The requested route does exist (so not a 404) but some part of the
        request is not acceptable, for example trying to access a resource by an
        ID that does not exist.
...
  schemas:
...
    InventoryItem:
      type: object
      description: Entity representing an instance of a stock of some product
      required:
        - productId
        - quantity
        - remainingQuantity
        - expires
      properties:
        id:
          type: integer
          example: 101
          readOnly: true
          description: >-
            A unique (within the inventory) identifer for this item -
            autogenerated by the system
        product:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/Product'
        productId:
          writeOnly: true
          type: string
          example: WATT-420-BEANS
          description: >-
            Product Code of corresponding product in business's catalogue.
            Notice that the ID is given when writing, but the full object is
            returned when reading
        quantity:
          type: integer
          example: 4
          description: How many items of stock are present
        remainingQuantity:
          type: integer
          example: 3
          description: How many items of stock are not in a listing
        pricePerItem:
          type: number
          format: double
          description: Price to purchase a single item of the product
          example: 6.5
        totalPrice:
          type: number
          format: double
          description: Price for purchasing the entire stock of the product
          example: 21.99
        manufactured:
          type: string
          format: date
          description: The date on which the product was prepared / manufactured
        sellBy:
          type: string
          format: date
          description: The date by which the product ought to be sold
        bestBefore:
          type: string
          format: date
          description: The "best before" date of the product
        expires:
          type: string
          format: date
          description: The expiry date of the product
    Image:
      type: object
      description: Entity representing an image stored on the server
      required:
        - filename
        - thumbnailFilename
      properties:
        id:
          type: integer
          example: 1234
          readOnly: true
          description: A unique identifier (system-wide) to represent this image
        filename:
          type: string
          example: /media/images/23987192387509-123908794328.png
          description: >-
            Relative URI to full size image. A fully qualified URI (e.g
            www.wasteless.com/media/images/...) should NOT be given.
        thumbnailFilename:
          type: string
          example: /media/images/23987192387509-123908794328_thumbnail.png
          description: >-
            Relative URI to thumbnail of image. A fully qualified URI (e.g
            www.wasteless.com/media/images/...) should NOT be given.
    Product:
      type: object
      description: Entity for catalogue items in a product catalogue
      required:
        - name
      properties:
        id:
          type: string
          example: WATT-420-BEANS
          description: >
            Developers can choose whether or not to make this a required field
            (i.e user  specified vs. auto-generated). In the case that the team
            automatically  generates the IDs, they should still allow this field
            to be present when  creating (POST) or modifying (PUT) a product,
            and then simply ignore its  value (i.e do not return a 400 code if
            it is present but not needed). An ID only needs to be unique within
            its catalogue, not system-wide.
        name:
          type: string
          description: The full name of the product
          example: Watties Baked Beans - 420g can
        description:
          type: string
          description: A desription of the product
          example: Baked Beans as they should be.
        manufacturer:
          type: string
          description: The manufacturer of the product
          example: Heinz Wattie's Limited
        recommendedRetailPrice:
          type: number
          format: double
          example: 2.2
          description: >-
            The recommended retail price of the product, given in whatever
            currency is in use at the business's location.
        created:
          type: string
          format: date-time
          readOnly: true
          description: >-
            The date and time when this product was created. This field is
            generated automatically by the server.
        images:
          type: array
          readOnly: true
          description: List of images of a product (implemented in U16).
          items:
            $ref: '#/components/schemas/Image'
...
    Page:
      type: object
      description: Page of results
      required:
        - count
      properties:
        count:
          type: integer
          example: 70
          readOnly: true
          description: Total number of results
...